Make PointerNativeTypeAssignmentRewriter non-nested
